@@ -10,16 +10,22 @@
         "logs": "firebase functions:log"
     },
     "engines": {
-        "node": "8"
+        "node": "10"
     },
     "dependencies": {
+        "@google-cloud/storage": "^5.0.1",
+        "@types/mime-types": "^2.1.0",
+        "@types/node": "^14.0.5",
         "cheerio": "^1.0.0-rc.3",
         "cors": "^2.8.5",
         "firebase-admin": "^8.6.0",
         "firebase-functions": "^3.3.0",
+        "fs-extra": "^9.0.0",
         "get-urls": "^9.2.0",
         "googleapis": "^51.0.0",
-        "node-fetch": "^2.6.0"
+        "mime-types": "^2.1.27",
+        "node-fetch": "^2.6.0",
+        "sharp": "^0.25.3"
     },
     "devDependencies": {
         "eslint": "^5.12.0",
