@@ -1,10 +1,13 @@
-import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
+import { Component, OnInit, ViewChild, ElementRef, ViewChildren, QueryList, HostListener } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { IonContent, IonList, Platform } from '@ionic/angular';
 
-import { Observable } from 'rxjs';
+import { Plugins } from '@capacitor/core';
+const { Clipboard, Browser } = Plugins;
+
+import { Observable, Subject } from 'rxjs';
 import { Subscription } from 'rxjs';
-import { switchMap, tap } from 'rxjs/operators'
+import { switchMap, tap, takeUntil } from 'rxjs/operators'
 
 import { AuthService, User } from '../../../../auth/shared/services/auth/auth.service';
 import { Profile } from '../../../../auth/shared/services/profile/profile.service';
@@ -12,6 +15,7 @@ import { TeamsService, Team } from '../../../shared/services/teams/teams.service
 import { MembersService, Member } from '../../../shared/services/members/members.service';
 
 import { Store } from 'src/store';
+import { MessageFileComponent } from 'src/app/shared/components/message-file/message-file.component';
 
 @Component({
   selector: 'app-direct',
@@ -19,7 +23,32 @@ import { Store } from 'src/store';
   styleUrls: ['./direct.component.scss'],
 })
 export class DirectComponent implements OnInit {
+  @HostListener('dragover', ['$event'])
+  onDragOver($event) {
+    $event.preventDefault();
+  }
+  @HostListener('dragleave', ['$event'])
+  onDragLeave($event) {
+    $event.preventDefault();
+  }
+  @HostListener('drop', ['$event'])
+  onDrop($event) {
+    $event.preventDefault();
+    const fileList: FileList = $event.dataTransfer.files;
+    console.log(fileList);
+    let files = [];
+    for (let i = 0; i < fileList.length; i++) {
+      files.push(fileList.item(i));
+      console.log(files);
+      if (files.length = fileList.length) {
+        console.log(files);
+        this.fileDropEvent(files);
+        files = [];
+      }
+    }
+  }
   finished = false;
+  @ViewChildren('childFile') childFiles:QueryList<MessageFileComponent>;
   @ViewChild(IonContent) contentArea: IonContent;
   @ViewChild(IonList, { read: ElementRef }) scroll: ElementRef;
   ios: boolean;
@@ -31,7 +60,6 @@ export class DirectComponent implements OnInit {
   team$: Observable<Team>;
   members$: Observable<Member[]>;
   member$: Observable<Member>;
-  newBody: string;
   directId: string;
   pathId: string;
   teamId: string;
@@ -44,6 +72,12 @@ export class DirectComponent implements OnInit {
   watch: boolean;
   member: Member;
 
+  newBody: string;
+  newFiles: any = [];
+  sending: boolean;
+
+  private readonly onDestroy = new Subject<void>();
+
   constructor(
     private store: Store,
     private activatedRoute: ActivatedRoute,
@@ -55,23 +89,26 @@ export class DirectComponent implements OnInit {
 
   ngAfterViewInit() {
     this.watch = true;
-    this.subscriptions = [this.memberSub]
-    this.memberSub = this.member$.pipe(tap(member => {
-      if (member && member.messages.length) {
-        this.member = member;
-        if (this.watch) {
-          this.scrollToBottom(0);
-          this.checkUnread();
-          this.mutationObserver = new MutationObserver((mutations) => {
-            this.scrollToBottom(500);
-            this.checkUnread();
-          })
-          this.mutationObserver.observe(this.scroll.nativeElement, {
-            childList: true
-          });
-        }
-      }
-    })).subscribe()
+    this.member$
+      .pipe(
+        takeUntil(this.onDestroy),
+        tap(member => {
+          if (member && member.messages.length) {
+            this.member = member;
+            if (this.watch) {
+              this.scrollToBottom(0);
+              this.checkUnread();
+              this.mutationObserver = new MutationObserver((mutations) => {
+                this.scrollToBottom(500);
+                this.checkUnread();
+              })
+              this.mutationObserver.observe(this.scroll.nativeElement, {
+                childList: true
+              });
+            }
+          }
+        })
+      ).subscribe()
   }
 
   checkUnread() {
@@ -93,6 +130,71 @@ export class DirectComponent implements OnInit {
     }
   }
 
+  async onPaste(ev) {
+    console.log(ev);
+    const result = await Clipboard.read();
+    console.log('Got', result.type, 'from clipboard:', result.value);
+    if (result.type !== 'text/plain') {
+      await this.newFiles.push(result);
+      return this.scrollOnFocus();
+    } else {
+      return false;
+    }
+  }
+
+  async fileRead(file) {
+    const reader = new FileReader();
+    return new Promise((resolve, reject) => {
+      console.log(file);
+      if (file.type === 'image/png' || file.type === 'image/jpeg' || file.type === 'image/jpg' || file.type === 'image/gif') {
+        reader.readAsDataURL(file);
+        reader.onerror = () => {
+          reader.abort();
+          reject(new DOMException("Problem parsing input file."))
+        }
+        reader.onload = () => {
+          resolve({ type: file.type, value: reader.result, name: file.name });
+        };
+      } else {
+        resolve({ type: file.type, value: file, name: file.name })
+      }
+    })
+  }
+
+  async fileDropEvent(files) {
+    const dropPromises = files.map(async file => {
+      try {
+        const result = await this.fileRead(file);
+        await this.newFiles.push(result);
+        return this.scrollOnFocus();
+      } catch (err) {
+        return console.log(err.message);
+      }
+    })
+    return Promise.all(dropPromises);
+  }
+
+  async fileChangeEvent(ev) {
+    const file = ev.target.files[0];
+    try {
+      const result = await this.fileRead(file);
+      console.log('res', result)
+      await this.newFiles.push(result);
+      return this.scrollOnFocus();
+    } catch (err) {
+      return console.log(err.message);
+    }
+  }
+
+  async removeFile(file) {
+    const index = this.newFiles.indexOf(file);
+    return this.newFiles.splice(index, 1);
+  }
+
+  async previewFile(message) {
+    await Browser.open({ url: message.body });
+  }
+
   scrollOnFocus() {
     setTimeout(() => {
       this.scrollToBottom(500);
@@ -105,9 +207,25 @@ export class DirectComponent implements OnInit {
     }
   }
 
-  sendMessage() {
-    this.membersService.addMessage(this.newBody, this.directId);
+  resetSender() {
+    this.newFiles = [];
     this.newBody = '';
+    this.sending = false;
+  }
+
+  async sendMessage() {
+    this.sending = true;
+    if (this.newFiles.length && this.childFiles.length) {
+      const uploadPromises = this.childFiles.map(async child => {
+        const upload = await child.upload();
+        return console.log(upload);
+      });
+      await Promise.all(uploadPromises);
+    }
+    if (this.newBody.length) {
+      await this.membersService.addMessage(this.newBody, this.directId, "message", null);
+    }
+    return this.resetSender();
   }
 
   onKeydown(event) {
@@ -131,6 +249,7 @@ export class DirectComponent implements OnInit {
       this.android = this.platform.is('android') && this.platform.is('capacitor');
       console.log(this.desktop, this.ios, this.android)
     })
+    
     this.date = new Date();
     this.time = this.date.getTime();
     this.newBody = '';
@@ -141,11 +260,6 @@ export class DirectComponent implements OnInit {
         tap(param => { this.directId = param.directId }),
         switchMap(param => this.membersService.getMember(param.directId))
       );
-    this.subscriptions = [
-      //this.authService.auth$.subscribe(),
-      //this.profileService.profile$.subscribe(),
-      //this.teamsService.teams$.subscribe()
-    ];
     this.team$ = this.activatedRoute.params
       .pipe(
         tap(param => { this.teamId = param.id }),
@@ -155,11 +269,7 @@ export class DirectComponent implements OnInit {
 
   ngOnDestroy() {
     this.watch = false;
-    this.subscriptions.forEach(sub => {
-      if (sub) {
-        sub.unsubscribe()
-      }
-    });
+    this.onDestroy.next();
   }
 
 }
